Implement Annotation support in KSP
This CL implements the AnnotationBox in KSP.

Unlike Java, we don't have 1 base class for elements in KSP.
Moreover, with synthetics, how we read annotations is not just
based on the KSAnnotated. For instance, when reading annotations
for a property, we should ignore all other annotations with
use sites other than field (or null, but that depends on the
declaration of the annotation).

To make the implementation manageable, I've created a new
XAnnotated interface for all annotation related methods and
KspAnnotated to implement them via delegates.
KspAnnotated also has built in filtering for use sites to
make them more declarative when specifying the delegate.

Unfortunately, KSP's java implementation has some known
bugs (https://github.com/google/ksp/issues/96) so I couldn't
add tests with java sources but I've added kotlin tests that
run with both kapt and ksp.
The same bug also prevents us from implementing use-site-inference
properly. When use-site is not declared, Kotlin uses the @Target
annotation on the annotation declaration, which has a list of
types hence cannot be read until #96 is fixed. For now, I've
implemented an approximation in KspAnnotated.

Bug: 160322705
Test: XAnnotationBoxTest
Change-Id: Ia4afffc0730ad346f46ac532e345d6006704c381