Remove initial owners from AndroidComposeView constructor
If they we not passed we got the values from ViewTree*Owners during the onAttachedToTheWindow calls. The whole idea of also passing owners as constructor params were to allow synchronous composition from Activity.onCreate() which is not the case anymore anyway as we changed the logic to first wait for ON_CREATED lifecycle even before composing to make sure it works correctly with owners from Fragments. With these changes there is no reasons to pass initial owners and duplicate the logic of fetching them.

Test: new test to make sure setContent function composes synchronously on an activity with CREATED state.
Change-Id: I8b751202b4d9291ef6f799bfd0482f963086ce83