Add NonDeclared types for KSP
In KSP, all types are modeled as KSType but that does not match
the model of Java's TypeMirror, DeclaredType and ArrayType.

This CL is an attempt to mimic the java types in KSP.

To do this, I've created sub classes for KspType for declared
and type arguments. The TypeArgument one is specifically necessary
to create the right typeName for them and also distinguish them
from declared types.

This CL is not fully complete, we need to make the KspTypeTests
run with KAPT to ensure we match KAPT as much as possible where it
makes sense (e.g. for nullability, KSP is better than KAPT and we
should keep it).
Similarly, we need to make all of XTypeTests run with KSP which
requires implementing primitive types (somehow) in KSP.
I'll implement those in followups.

Bug: 160322705
Test: XTypeTest, KSTypeExtTest, KspTypeTest

Change-Id: Icd0e91ca14ce9fd34924e5c4cb90b7ac26aaff84