Fix remeasuring after LayoutNode removal
Fixes: 170287733

Relnote: N/A

After a LayoutNode is removed, it is possible to still
receive events for that node. This CL adds two factors
to avoid operating on removed LayoutNodes. First, it
checks isValid on all scopes in the callbacks. Second,
it uses postAtFrontOfQueue() to clear the listeners
for removed LayoutNodes as quickly as possible.

Test: new test, manual testing of breaking code
Change-Id: I3909fd3a5dbba3a4b85211743b652ad2ecda418d