Fix the preview is stretched on Nexus 4
There is a Camera1/HAL1 issue on the Nexus 4. The preview will be stretched when configuring a JPEG that doesn't actually have the same aspect ratio as the maximum JPEG resolution.

Workaround the issue by only configuring the size has the same aspect ratio as the maximum JPEG resolution to all the UseCases on Nexus 4.

Relnote:"Fixed an issue where the preview was stretched on Nexus 4 devices running Android L (API levels 21 and 22)."

Bug: b/158749159
Test: run TargetAspectRatioTest
Change-Id: I4d407ee366c978de4e518995e97d5946fc19e29a