Add ability to get TypeName from JVM signature
This CL adds a new method to JvmDescriptorUtils to create a
TypeName from a signature.

It is not used now but will be used in a followup when we
start creating TypeNames in Java realm.

Also noticed that we were not running jvm descriptor utils
tests at all :D, fixed.

Bug: 160322705
Test: JvmDescriptorUtilsTest
Change-Id: I8d6903879c4021b461282f70b8172f0237798ed9