Make the SharedPreferencesView constructor internal.
Originally we made this constructor public for testing. But I think the best way to test is to populate a SharedPreferences and run the SharedPreferencesMigration.

Bug: 159847353
Test: All existing tests pass
Relnote: Making the constructor for SharedPreferencesView internal. It was originally public to allow for testing. Tests should instead construct a SharedPreferencesMigration and test against that.
Change-Id: I938912884b2f088e70262fecc9c81a80795e1579