Fix NPE of findContainingItemView in clearFocus
If all of the following conditions are satisfied:
1. device is P and above
2. RecyclerView is a child of ScrollView
3. notifyItemChange with payload removes focus from currently selected
item.

clearFocus() will call rootView.requestFocus, which is a top-down
process. ScrollView requestFocus() calls RecyclerView.addFocusables(),
at this time even though the focus flag was cleared from the itemView,
RecyclerView is in *uncertain* state: hasFocus() is still true, but
findFocus() is unexpectedly null.

So we cannot assume findFocus() is not null even hasFocus() is true.
Fix is to add a null check.

Bug: 161359022
Bug: 156578022
Test: ./gradlew leanback:leanback:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.leanback.widget.GridWidgetTest#testBug161359022

Change-Id: Id0e42555cfe720eb291f0c80dc3dab340845783e