Allow arbitrary "extra" in version definition.
It is needed for publishing MPP artifacts from JetBrains CI.

Very often we build versions not from android-master-dev (because we need CL's that not merged yet).

And it is would be very helpful if we can name version as "0.0.1-scrollbars-build1".

Also it may be used for versions like "0.0.1-m1" (milestone 1)

Test: ./gradlew buildOnServer
Change-Id: I8c7dba43f9dfb94ff4c16479210e31258f235997