Add KeyboardOptions configuration class
This CL adds KeyboardOptions that is very similar to IMEOptions
except that it doesn’t have the singleLine option. The signal
of singleLine that is sent to IME is different that the
implementation of TextFields that will provide the single line
functionality with scroller, softwrap etc.

KeyboardOptions is for TextFields to use whereas IMEOption
is for communicating with TextInputService.

Will add KeyboardOption to the following TextFields in a
separate change:
- BasicTextField
- OutLinedTextField
- TextField

Test: ./gradlew compose:foundation:foundation:test
Test: ./gradlew compose:foundation:foundation:connectedAndroidTest

RelNote: Added KeyboardOptions for use in TextFields

Bug: 171022145
Change-Id: I9ca329336d80be07e6a0c1e6d7a2f84902774898