Remove undisambiguatable kdoc link for insertSeparators
From the official docs: https://kotlinlang.org/docs/reference/kotlin-doc.html

"Note that KDoc does not have any syntax for resolving overloaded members
in links. Since the Kotlin documentation generation tool puts the
documentation for all overloads of a function on the same page,
identifying a specific overloaded function is not required for the link
to work."

It looks to be impossible to disambiguate kdocs links, so our only
recourse is to remove + describe what we actually mean here.

Fixes: 171830512
Test: N/A kdocs change only
Change-Id: If455ad439da7613934d76aad2564d31f5edc6082