Remove navigation graphs without destinations
When using navigate() with popUpTo, you may pop
the last destination of the current graph. In those
cases, the graph itself should be popped off the
back stack before the new destination (and optionally
any hierarchy associated with it) is added to the
back stack.

Relnote: "Fixed an issue where using `navigate()` with
a `popUpTo` that removed the last destination in a
navigation graph would not immediately destroy and
remove the navigation graph itself from the back stack."
BUG: 171389496
Test: new NavBackStackEntryLifecycleTest test

Change-Id: I910a3aadd9add3932285cc9aefbb4cfc45ec178b