Ensure that navigating to a graph creates a new graph
Given a navigation graph of:
A ->
  B ->
    C
  D

If you are at C and call navigate(B), we should
create a new copy of B (and its start destination of C)
rather than reusing the existing instance of B. This
differentiates the case of navigate(B) (which would make the
back stack transition from A>B>C to A>B>C>B*>C*) from
navigate(C) (which would make the back stack transition
from A>B>C to A>B>C>C*).

Special consider was also taken into account for
orphaned graph to ensure that they are popped correctly
before the new copies of the graph are added to the
back stack.

At a practical level, this ensures that retrieving the
NavBackStackEntry of the parent graph has the correct
arguments.

Relnote: "Fixed an issue where navigating to nested
navigation graph would not create a new graph instance
on the back stack."
BUG: 171389496
Test: new NavBackStackEntryLifecycleTest tests

Change-Id: Ifc8318d58fedcfe71ed9a486fe43f963ebb4a330