Suppress inspector info modifier check for inline modifiers
Suppress the lint checks for:
- Modifier.onGloballyPositioned
- Modifier.onSizeChanged

In both cases these modifiers are declared inline with a lambda
parameter declared crossinline. It is not allowed to access the
lambda in the inspector information beacuse it is "crossinlined".

I am not sure why these modifiers are setup this way.
It seems to me that the materialized modifer will always allocate
the lambda so it could be declared "noline".

Bug: 172375389
Test: N/A
Change-Id: I0b08911fbe936ffafcb4f4c017148054791b8f97