Add an API to check if the window is in focus
The current Focus modifiers let you know if a composable is in focus. When there are multiple windows open on the screen, we need to provide some API that can be used to check if the current window is in focus. This CL adds WindowManager.isWindowFocused property and a WindowFocusObserver composable that provides an onWindowFocusChanged callback that can be used to check if the window is in focus.

Bug: 170932874
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.platform.WindowManagerAmbientTest
Relnote: Added WindowManager.isWindowFocused to check if the host window is in focus, and a WindowFocusObserver that provides an onWindowFocusChanged callback.
Change-Id: I53b2a702b81215dc5a5536144a752e1c93ab056e