InteractiveWatchFaceWcsClient to use ComplicationState
Also changes InteractiveInstanceManager.releaseInstance to silently
ignore an unknown instanceId.  This is necessary because when
running locally the golden image tests don't work and the
instance gets deleted before the release call happens leading to
an exception and test flakiness on bots.  I think in practice we can't
usefully recover from this and we don't want to crash the watch
face.

Test: Presubmit
Relnote: Fix InteractiveWatchFaceWcsClient to use ComplicationState insted of the wire format.
Change-Id: Icb8a49513e834ead46c7bcca78504f905b8766ba