Fix canSupportGuaranteedSizeBack/Front tests failures on OnePlus 7T
The original supported sizes grouping logic used for-loop and hasMatchingAspectRatio() to find a key from the HashMap key set. The for-loop would break when any key was found to match the size under mod16 matching condition. In the issue, the for-loop stopped at key of aspect ratio 512/369 when grouping 640x480. But 640x480 should be grouped into 4/3 size group so that it can be found when the target resolution setting is 640x480.

To fix this issue, the following changes are made:
1. When grouping the supported sizes by aspect ratio, adding a size to all groups that can be matched by hasMatchingAspectRatio().
2. After sorting the groups by aspect ratio distance and collect all sizes into the final result list, keeping only one in the final result list. The finally kept item belongs to the group that has the minimal aspect ratio distance to the target aspect ratio.

Bug: 172394922
Test: ./gradlew camera:camera-camera2:build && ./gradlew camera:camera-camera2:connectedAndroidTest
Change-Id: I740120106f59c0734ad952d0b489921e5781e238