Ensure interrupted animations are handled properly
In aosp/1442517 an issue with interrupting exit animations was
addressed by ensuring the animations were readded to the container
between before attempting to start the enter animation.

Doing that as part of the ACTIVITY_CREATED state in the
FragmentStateManager was too early since it is possible that the
animation had not yet actually been canceled. So instead we moved it
right before the enter animation starts as part of the onStart()
callback in the SpecialEffects Operation. This ensures we the add the
view at the last possible moment before starting the enter animation and
it cannot be removed by some other call.

Test: FragmentAnimationTest.popReplaceOperationWithAnimations
Bug: 171745757
Change-Id: I9533ad9c51bbe2a44c0c46770c73453f988ec2ab