Slight improvements in property modifiers
This CL fixes a bug in how we calculate modifiers for the fields
derived from a property.

Previously, we checked if it is JvmField but it does not cover
java sources. Unfortunately, for .class files, we don't have
enough information as whether we should read modifiers from a]
property or not depends whether it was compiled from a java or
kotlin source. KSP does not provide the kotlin metadata annotation
hence we cannot check whether the .class files was compiled from
java or kotlin source. As an approximation, for class files, we
check whether it has the jvm field annotation or if it is delegated.

This is unlikely to be correct but passes tests :) so better than
before.

It also hits a small visibility issues w/ KSP for package private
fields. https://github.com/google/ksp/issues/159
I just added a link in the test for now as it is unlikely to effect
Room's use cases.

Last but not least, to test .class files, i had to add new infra to
pre-compile a set of classes to add as a dependency. I'm leaning
towards extracting all this testing infra into another module (from
compiler processing's test code) so that Room can start using it too
when it wants to run compilation tests with KSP.

Bug: 160322705
Test: KspFieldElementTest
Change-Id: I738e86fa24cdeaf2be8e91418a1b2a3e0a05f903