Add XType.makeNullable / makeNonNull
There are situations where room grabs types in hardcoded
places, e.g. we want to get the List type.
In those situations, the nullability of that type is unknown
as it is not obtained from source.
This is an issue when room means to mark them as nullable, for
instance in cursor value readers/writers where room uses boxed
types to denote nullability which is not enough as in KSP, they
just look like non-null kotlin.Int.

This CL introduces makeNullable/makeNonNull to make those cases
clear. There are certain cases where changing nullability of a type
might also change the type. For instance, primitives cannot be
marked as nullable. In these cases, it gets automatically boxed.
Note that it is not a two way conversion as it is totally valid to
mark a boxed primitive as non-null.

I've also noticed that we were not boxing `void` type properly
in javac processor, fixed that as well.

Bug: 160322705
Test: XNullabilityTest
Change-Id: Iad59aa7520c0995de05e8160b02c745113242e61