Change default anchor to center. Simplify code.
Changing the WearCurvedTextView's default anchorType to center, means we
can simplify the contract between the arc layout and the curved text,
the curved text by default draws itself x-centered at the top, and the
parent arc layout can rotate it to where it needs to be.

Relnote: Change WearCurvedTextView's anchor type default to center.
Test: Updated

Change-Id: I105ff40f39488f960582a7cba799e47b6b7086ee