Handle variance when overriding methods
For generics, the code kotlin would generate depends on the variance
of the parameter or its inheritance chain.

This CL implements the same algorithm so that room can generate the
correct overrides.

This is only relevant when generating overrides in Room hence it is
tested in MethodSpecHelperTest.

I've also noticed that when a suspend function has varargs, kotlin
actually generates an array since last argument has to be a
continuation. Added a check to avoid returning such methods as vararg.

I've changed continuation parameter name to match kapt as
well so that tests can do string comparison. It only helps with golden
testing, not really an important change binary wise.

Bug: 160322705
Test: MethodSpecHelperTest, XExecutableElementTest
Change-Id: I1a53a4513b2ac772332246e979cf86ad846c878d