Improve detectMultitouchGestures API and demo
This adds a centroid parameter to the detectMultitouchGestures callback
and uses it in the demo to make zoom and rotate work as they do in
Google Maps: that is, the centroid is the stable anchor that the rest
zooms/rotates around.

Also, I changed the callbacks to be one single callback taking all the
parameters.  This makes it easier to have a polished, performant
implementation in use cases that need all three, and it's easy to use a
wrapper to convert it to a 3-callback style if desired (whereas merging
3 callbacks into 1 is nontrivial).

Test: MultitouchGestureDetectorTest
Relnote: "detectMultitouchGestures now uses one callback with combined
centroid, pan, zoom and rotate parameters."

Change-Id: Ie6e1c2d7496ed0aabc6a5256f859c91cc39a08dc