New processing test APIs
This CL updates the compilation testing infra and refactors the
internals.
The biggest change is that now there is an API to assert compilation
results rather than calling `runXXXforFailed` methods.
In this new API, the handler is responsible to make a failure assertion
if the compilation should fail, otherwise, test will expect compilation
to succeed and fail the test if the compilation fails.

As we add tests in Room, I'll add APIs to also assert generated code but
right now we don't have that use case hence not adding them (also, it is
not clear yet how it will work with kotlin compile testing).

To be able to collect diagnostics before it goes to the compilation
infra, I've moved XMessager to an abstract class (from interface) and
add ability to watch messages. Unfortunately, kotlin compile testing
does not provide an API to get messages with diagnostic types which is
why we have the abstraction on our end.

Bug: 160322705
Test: existing tests
Change-Id: I0463b1bdc822e3e1e7fd39f3e5e17581c3bfde19