Clean up coroutines in repeated tests like benchmarks
Although composition.dispose() causes all coroutines in the composition
to get cancelled (terminate with an exception), their continuations were
still holding onto memory in the benchmark environment.  That's because
the shared TestCoroutineDispatcher was still holding a reference to them
in its pending queue.  This pokes the dispatcher to flush its queue
after dispose.

Test: MemoryLeakTest

Change-Id: I32dc27e18815939a5c7e9e4f4cfcaf36025b8555