listChoiceBackgroundIndicator documentation
It is a drawable used as a background for activated list items, not selected ones. There is a difference between a selected and an activated state as stated in Android's documentation. A choice in a list represents a non-transient active state, as opposed to, for example, a transient selection of files in a file manager's list (the one which triggers an ActionMode). This mistake stems from the AOSP itself. I will make a pull request with the same fix there too.